Add bash function to re-use cloned remote repositories
Git remote repositories are save as bare repositories in

downloads/repos

and the build is done in a locally cloned repository in

downloads/builds

using the "--shared" option.

When a build in requested again the local bare repository is
reused and only the build repository is recreated.

Avoid fetching over and over again the same repository from the
internet.