Improve generated environment script
Distinguish between pkgconfig path and dir. The latter is a single
directory used to install pkgconfig file during libraries install.

Now define the "_prefix" dir for the environment's directory at the top of
the script and use "${_prefix}" through the rest of the script.

Use also a one-liner approach to add a path to environment variable:

export PATH="/additional/path${PATH:+:}${PATH}"

so that the colon is added only when PATH is actually defined and not
empty.