Fix DESTDIR problem on Windows
Now DESTDIR is no longer used. Instead the initial PREFIX using when
building is the temporary directory used to create the package. Then
the name of the temporary directory is replaced with
__LHELPER_PREFIX__ (is relocated).

Later, when extracting the package, the string __LHELPER_PREFIX__ is
replaced with the new PREFIX.

To improve: the SED_PREFIX_PATTERN transform the install prefix in a
SED regex patten except the '/' that are not touched but are later
escaped by the function library_dir_reloc. The work of transforming
the prefix path in a SED regex should be done all at once at the
beginning.