Do not use packaged file when version matches git-*
Useful to compile package from a git branch. The version used is git-*
like for example git-master or git-dev. The recipe is selected using
the usual rules. The package is created but when the same library is
installed again the existing package is ignored and the library is built
again.