Add '--pic' option for install command
It is introduced because on Linux, if a library is compiled as a static
library and it is later used to build a shared library the link will
fail.

The reason is that configure and cmake does not add the -fPIC flag if
the shared library is disabled.

On the other hand Meson always add the -fPIC flag so for Meson this is
either unneeded or the logic should be inverted: ask Meson to remove
-fPIC if the --pic option is not given.