Use source command to execute the package recipe
The recipe was executed using "bash -e" so a new bash shell was started to
execute the recipe. A command "export -f" was used to export the functions
needed by the recipe.

Now we use instead the "source" command to execute the recipe from the
existing bash process and we remove the export command which is no longer
needed.