Add a mechanism to declare dependencies
Let the recipe declare its dependencies and store them
in a file.

The list of dependencies is stored but is currenly not used.

The recipe is run a first time to store only the dependecies
using a flag that prevent downloading the source and building.
The second time is run to actually build the library.