Experimental setup using Lua and reproc library
Initial setup to compile an executable running arbitrary
Lua code with the additional module "process" using the
"reproc" library to start child processes.

The idea is that with this setup we can use Lua for all the
logic for lhelper and start processes using the "reproc"
library with the "process" module.