Add use directive for recipes
With the use directive, 'declare_use', recipes declares that they
can use a library if present but it is not declared.

The presence of the declared library is not enforced but we code into
the package digest the version of the library actually present of if
none is present.

Fix other bash madness and errors.