Move environments into project's directory
No longer use the absolute directory $prefix/var/lhenv but use
instead .lhelper/env inside the project's directory.

Introduce the new environment variable

LHELPER_ENV_ROOT

that points to the projects directory. It is used to locate the .lhelper
file when the "edit" command is used.

Now LHELPER_ENV_PREFIX / ROOT are both stored in the "activate" script
as relative paths and are resolved only when the environment is
activated.