Address environments by their names
No longer use numbers for environments but address them by their
name.

To find if an environment match looks only to its config, not to the
list of packages. If it matches the installed packages are updated.
If it doesn't match remove the previous environment and create a new
one.