Allow build files without CPU variables
Now it is possible and even recommended to use generic build files
that does not specify the CPU_TYPE and CPU_TARGET variables.

When activating the environment it will use the environment variables
set at the moment of the activation and if not set it will guess based
on the architecture using a conservative approach for CPU_TARGET.

In this way a single build file can work for any architecture but it
is still possible to specialize the build file for a specific
architecture.