Fix serious problem with subshell reload
Fixed the problem with the build filename for the "edit" command. Now
the build filename is stored in an environment variable relative to the
working directory.

Add explicitly in the activate script the name of the project directory
so that if the directory is relocated the packages are reinstalled.

Fixed a serious problem with the live reload of the environments. The
problem was that sourcing of the environment activating script was made
in the parent bash process and they were cumulated at each restart. Now
the activate script of the environment is sourced only withing a bash
subshell used to activate the environment. As a consequence the
start_subshell function use exec now to remain in the activation
subshell.