Fix logic error in lh-path-replace
The case when a prefix in find_replace_prefix_path is found but the
following pattern does not was incorectly handled. Now it will not end
the search but will just search a little further in the content.

A similar problem was fixed in find_replace_path but this latter
modification has not been tested.