Fix realpath implementation
Fix the linux version to use realpath instead of readlink.

When a file does not exist lookup for its directory and if it exists
provide a realpath based on the realpath of the directory. This behavior
reproduce the behavior of the realpath command on linux and is more
useful in general.