Fix pattern matching logic in lh-path-replace
When deciding if a string part match the pattern we check now that the
character before the matching part is not a path-like character.

The goal of this new check is avoiding to replace pattern occurrences
that are a sub-path of a longer path. For example is the pattern to
replace is /usr we will not match its occurrence in /something/foo/usr
because it is a subpath of a wider path.

We want also avoid replacing /usr in C:/msys64/usr like it can occour on
windows using MSYS2.