Manage prefix options when doing meson build
It is now possible when doing a meson build to pass a prefix option. If
present it will inhibit lhelper from providing its own prefix option
when doing the build but it will use instead its own working prefix when
installing as a --destdir option.

This option is meant for packages that really needs to be configured to
have a fixed, system-dependent prefix. In this case the package recipe
is supposed to adjust the paths that needs to be adjusted from the
recipe, after the meson build has run.