feat: package build with true system's prefix
It works, at least on linux with a test on a few libraries EXCEPT in
some cases it replace some occurrences that match $prefix but they
should not be replaced at all because they are hard-coded paths.

The problem happens because it will replace in all text files a short
generic path like "/usr" with a custom lhelper's string but a path like
/usr can naturally occurs in places unrelated to $prefix where it should
not be replaced.

For example it happens with glib for some python script it installs as
tools. The first line is a "#!/usr/bin/env python3" and the /usr path is
getting replaced but that should not happen.