fix: several problem related to prefix and destdir
Now recipes are mostly working but more testing is needed. Main problem
is that we choose which files to relocate, .pc, .la and -config files but
this list may be incomplete and some libraries may need to fix other files.

Probably we should give the opportunity to each recipe to relocate the
files they choose after the "build_and_install" command.

We may also use the meson option -Dpkgconfig.relocatable=true.

The prefix stuff is not fixed on windows and a problem related to packages
to destdir was solved. The problem was that with destdir meson was installing
file in $destdir/$prefix where we wanted instead them in $destdir.
In addition on windows meson and cmake transform /usr to C:/msys64/usr and
on install files are located into $destdir/msys64/usr.