fix: problem with environment std directories
We make sure the environment's standard include and lib directories are
included since the beginning when using the configure script. In this
way when configure search for a library using the standard include path
the default directories of the environment are taken into account.

In principle if configure was using pkg-config each time that should not
be needed but this is not always the case. For example libpng16 search
for zlib ignoring the pkg-config file, it just search the header file
and the library itself in the standard directories as defined by the
compiler.

This fix works but there is a function that is duplicate in the code.